+++ packages/parser/src/language/architecture/tokenBuilder.ts
@@ -1,0 +1,7 @@
+import { AbstractMermaidTokenBuilder } from '../common/index.js';
+
+export class ArchitectureTokenBuilder extends AbstractMermaidTokenBuilder {
+  public constructor() {
+    super(['architecture']);
+  }
+}