+++ packages/parser/src/language/index.ts
@@ -5,6 +5,7 @@ export {
   PacketBlock,
   Pie,
   PieSection,
+  Architecture,
   GitGraph,
   Branch,
   Commit,
@@ -16,16 +17,19 @@ export {
   isPacketBlock,
   isPie,
   isPieSection,
+  isArchitecture,
   isGitGraph,
   isBranch,
   isCommit,
   isMerge,
 } from './generated/ast.js';
+
 export {
   InfoGeneratedModule,
   MermaidGeneratedSharedModule,
   PacketGeneratedModule,
   PieGeneratedModule,
+  ArchitectureGeneratedModule,
   GitGraphGeneratedModule,
 } from './generated/module.js';
 
@@ -34,3 +38,4 @@ export * from './common/index.js';
 export * from './info/index.js';
 export * from './packet/index.js';
 export * from './pie/index.js';
+export * from './architecture/index.js';