+++ packages/parser/src/parse.ts
@@ -1,8 +1,8 @@
 import type { LangiumParser, ParseResult } from 'langium';
 
-import type { Info, Packet, Pie, GitGraph } from './index.js';
+import type { Info, Packet, Pie, Architecture, GitGraph } from './index.js';
 
-export type DiagramAST = Info | Packet | Pie | GitGraph;
+export type DiagramAST = Info | Packet | Pie | Architecture | GitGraph;
 
 const parsers: Record<string, LangiumParser> = {};
 const initializers = {
@@ -21,6 +21,11 @@ const initializers = {
     const parser = createPieServices().Pie.parser.LangiumParser;
     parsers.pie = parser;
   },
+  architecture: async () => {
+    const { createArchitectureServices } = await import('./language/architecture/index.js');
+    const parser = createArchitectureServices().Architecture.parser.LangiumParser;
+    parsers.architecture = parser;
+  },
   gitGraph: async () => {
     const { createGitGraphServices } = await import('./language/gitGraph/index.js');
     const parser = createGitGraphServices().GitGraph.parser.LangiumParser;
@@ -31,6 +36,7 @@ const initializers = {
 export async function parse(diagramType: 'info', text: string): Promise<Info>;
 export async function parse(diagramType: 'packet', text: string): Promise<Packet>;
 export async function parse(diagramType: 'pie', text: string): Promise<Pie>;
+export async function parse(diagramType: 'architecture', text: string): Promise<Architecture>;
 export async function parse(diagramType: 'gitGraph', text: string): Promise<GitGraph>;
 
 export async function parse<T extends DiagramAST>(